blueprint builder: move external networking choices to submodule
