+++ nexus/reconfigurator/planning/src/blueprint_builder/mod.rs
@@ -5,6 +5,7 @@
 //! Low-level facility for generating Blueprints
 
 mod builder;
+mod external_networking;
 mod zones;
 
 pub use builder::*;
